refactor: move kubernetes runtime mock code to separate file
More mock code will be forthcoming to facilitate tests in other packages.
